Module rustysensor::em

The Electromagnetics module

Modules

consts

tables

Functions

absorption_length

Computes absorption length

angstroem_attenuation

Computes the Angström attenuation given the base attenuation

angular_frequency

Computes angular frequency from regular frequency

bb_radiation

Computes total black body radiance Formula: $radiation = \sigma T^4$

doppler_ratio

Computes Doppler ratio given velocity of system and angle at system

electric_permeability

Computes $\epsilon = \epsilon_r\epsilon_0$

em_frequency

Computes the electromagnetic frequency from wavelength

em_wavelength

Computes the electromagnetic wavelength given regular frequency

exit_angle

Using Snell’s law, computes the exit angle of an entrant ray of light given two refracting indexes and an entrance angle. Snell’s law is given by $n_1 \sin(\theta_1) = n_2 \sin(\theta_2)$

flux_density

Computes flux density of E field given its amplitude

fog_liquid_mass_density

Computes the liquid mass density of fog

fog_scattering_coefficient

Computes the scattering coefficient of fog.

gas_dielectric_constant

Computes the dielectric constant of a gas

gas_refractive_index

Computes the refractive index of a gas

homogeneous_material_eb_ratio

Gets the ratio of the magnitudes of (90 degree phase) E field vs B field in homogenous materials

irradiance

Computes the irradiance using a lambda which has either the incoming or outgoing L values. If Lincoming, returns the irradiance, if Loutgoing then returns the radiant exitance

magnetic_permeability

Computes $\mu = \mu\mu_0$

metal_absorption_length

At radio frequencies, we can use this approximation to compute the absorption length for metals given conductivity

metal_dielectric_imag

Computes the imaginary part of $\epsilon$ for a metal

metal_dielectric_real

Computes the real part of $\epsilon$ for a metal

photon_energy

Computes photon energy given frequency

plasma_dielectric_constant

Computes the dielectric constant of a plasma

plasma_phase_velocity

Computes the phase velocity of a plasma. NOTE: Phase velocities may be higher than the speed of light

plasma_refractive_index

Computes the refractive index of a plasma

plasma_transparency_frequency

Computes the EM frequency at which a plasma becomes transparent

refractive_index

Computes refractive index given $\epsilon_r$ and $\mu_r$.

spectral_radiance_f

Computes $L_f$ (the spectral radiance) using the Rayleigh-Jeans approximation It is computed using the following formula: $$L_f = 2K\frac{T}{\lambda^2}$$

spectral_radiance_lambda

Computes $L_lambda$ using Rayleigh-Jeans approximation It is computed using the following formula: $$L_\lambda = 2K\frac{Tc}{\lambda^2}$$

wave_num

Computes wave number given wavelength

windowed_fraunhofer_diffraction

Fraunhofer diffraction Note: Fraunhofer diffraction generally just requires fft For some window, w, we can just use this $sinc()$ approximation: $diffraction = \frac{n w \sin(\theta)}{2}$